Move all preloading code into preload package
