+++ pkg/drivers/kic/oci/volumes.go
@@ -21,12 +21,11 @@ import (
 	"bytes"
 	"fmt"
 	"os/exec"
-	"path"
 	"strings"
 
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
-	"k8s.io/minikube/pkg/minikube/localpath"
+	"k8s.io/minikube/pkg/drivers/kic/preload"
 )
 
 // DeleteAllVolumesByLabel deletes all volumes that have a specific label
@@ -91,7 +90,7 @@ func allVolumesByLabel(ociBin string, label string) ([]string, error) {
 }
 
 // CreatePreloadedImagesVolume creates a volume with preloaded images
-func CreatePreloadedImagesVolume(k8sVersion string) (string, error) {
+func CreatePreloadedImagesVolume(k8sVersion, baseImage string) (string, error) {
 	if err := PointToHostDockerDaemon(); err != nil {
 		return "", errors.Wrap(err, "point host docker-daemon")
 	}
@@ -102,10 +101,9 @@ func CreatePreloadedImagesVolume(k8sVersion string) (string, error) {
 	if err := createDockerVolume(volumeName); err != nil {
 		return "", errors.Wrap(err, "creating docker volume")
 	}
-	targetDir := localpath.MakeMiniPath("cache", "preloaded-tarball")
-	tarballPath := path.Join(targetDir, fmt.Sprintf("%s.tar", k8sVersion))
+	tarballPath := preload.TarballPath(k8sVersion)
 
-	if err := extractTarballToVolume(tarballPath, volumeName); err != nil {
+	if err := extractTarballToVolume(tarballPath, volumeName, baseImage); err != nil {
 		return "", errors.Wrap(err, "extracting tarball to volume")
 	}
 	return volumeName, nil
@@ -129,11 +127,11 @@ func dockerVolumeExists(name string) bool {
 	return false
 }
 
-func extractTarballToVolume(tarballPath, volumeName string) error {
+func extractTarballToVolume(tarballPath, volumeName, imageName string) error {
 	if err := PointToHostDockerDaemon(); err != nil {
 		return errors.Wrap(err, "point host docker-daemon")
 	}
-	cmd := exec.Command(Docker, "run", "--rm", "-v", fmt.Sprintf("%s:/preloaded.tar:ro", tarballPath), "-v", fmt.Sprintf("%s:/extractDir", volumeName), "busybox", "tar", "xvf", "/preloaded.tar", "-C", "/extractDir")
+	cmd := exec.Command(Docker, "run", "--rm", "--entrypoint", "/bin/bash", "-v", fmt.Sprintf("%s:/preloaded.tar:ro", tarballPath), "-v", fmt.Sprintf("%s:/extractDir", volumeName), imageName, "tar", "-I", "lz4", "-xvf", "/preloaded.tar", "-C", "/extractDir")
 	fmt.Println(cmd.Args)
 	if out, err := cmd.CombinedOutput(); err != nil {
 		return errors.Wrapf(err, "output %s", string(out))
